\name{colortools-package}
\alias{colortools}
\docType{package}
\title{
  Tools for colors in an HSV color model
}
\description{
  Handy functions to help users select and play with color schemes in an HSV color model
}
\details{
\tabular{ll}{
	Package: \tab colortools\cr
	Type: \tab Package\cr
	Version: \tab 0.1-0\cr
	Date: \tab 2012-06-13\cr
	License: \tab GPL (>=3)\cr
}
  The package provides several functions to get color wheels, as well as other utilities to choose color schemes in an HSV color model 
}
\author{
  Author: Gaston Sanchez

  Maintainer: Gaston Sanchez <gaston.sanchez@berkeley.edu>
}
\keyword{ package }

